requirements.txt
google-generativeai==0.3.2
google-search-results==2.4.2
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
gspread==6.0.2
oauth2client==4.1.3
